+++ tests/test_processor.py
@@ -143,30 +143,30 @@ class NifiProcessor(unittest.TestCase):
     #         assert str(e) == 'Failed to start processor'
     #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUNNING')
     #
-    # @patch('NifiLibrary.NifiLibrary.NifiLibrary.update_process_state')
-    # def test_run_once_processor_runs_successfully(self, mock_update_process_state):
-    #     mock_update_process_state.return_value = 'Success'
+    # # @patch('NifiLibrary.NifiLibrary.NifiLibrary.update_process_state')
+    # # def test_run_once_processor_runs_successfully(self, mock_update_process_state):
+    # #     mock_update_process_state.return_value = 'Success'
+    # #
+    # #     result = self.nifi.run_once_processor(self.processor_id, return_response=True)
+    # #
+    # #     assert result == 'Success'
+    # #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
     #
-    #     result = self.nifi.run_once_processor(self.processor_id, return_response=True)
-    #
-    #     assert result == 'Success'
-    #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
-
-    def test_run_once_processor_raises_exception_for_missing_processor_id(self):
-        try:
-            self.nifi.run_once_processor(None)
-        except Exception as e:
-            assert str(e) == 'Require parameters cannot be none'
-
-    # @patch('NifiLibrary.NifiLibrary.NifiLibrary.update_process_state')
-    # def test_run_once_processor_logs_error_on_failure(self, mock_update_process_state):
-    #     mock_update_process_state.side_effect = Exception('Failed to run processor once')
-    #
+    # def test_run_once_processor_raises_exception_for_missing_processor_id(self):
     #     try:
-    #         self.nifi.run_once_processor(self.processor_id)
+    #         self.nifi.run_once_processor(None)
     #     except Exception as e:
-    #         assert str(e) == 'Failed to run processor once'
-    #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
+    #         assert str(e) == 'Require parameters cannot be none'
     #
+    # # @patch('NifiLibrary.NifiLibrary.NifiLibrary.update_process_state')
+    # # def test_run_once_processor_logs_error_on_failure(self, mock_update_process_state):
+    # #     mock_update_process_state.side_effect = Exception('Failed to run processor once')
+    # #
+    # #     try:
+    # #         self.nifi.run_once_processor(self.processor_id)
+    # #     except Exception as e:
+    # #         assert str(e) == 'Failed to run processor once'
+    # #     mock_update_process_state.assert_called_once_with(self.processor_id, 'RUN_ONCE')
+    # #
     if __name__ == '__main__':
         unittest.main()