+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
+## [2.1.2] - 2024-08-17
+### Changed
+- Updated keywords to support setting the NiFi endpoint. 
+- Comment run_once_processor function because Nipyapi does not allow RUN_ONCE state.
 ## [2.1.0] - 2024-08-17
 ### Added
 - added new keyword `Set access token` to set the access token for the NiFi instance.